🎨 Format

+++ src/project/routes/admin.py
@@ -1,7 +1,7 @@
 import re
 from datetime import datetime
 
-from flask import Blueprint, redirect, render_template, request, url_for
+from flask import Blueprint, abort, redirect, render_template, request, url_for
 
 from ..extensions import db
 from ..models import Industry, InvestmentFirm, Investor, Round, User, UserInfo, UserPayment
@@ -12,7 +12,6 @@ from ..utils.errors.auth_error_messages import (
     AUTH_MISMATCHED_PASSWORDS,
     AUTH_OAUTH_USED,
 )
-from ..utils.google_storage import prepare_picture, upload_blob
 from ..utils.info_lists import languages as language_list
 from ..utils.status_enum import OauthProvider, Status, StatusType
 
@@ -123,8 +122,6 @@ def validate_field(value, field_name, error_message, redirect_url):
     if value != value.strip():
         status = Status(StatusType.ERROR, f"{field_name} cannot start or end with spaces.").get_status()
         return redirect(url_for(redirect_url, _external=False, **status))
-    return None
-
 
 
 # investors
@@ -149,25 +146,27 @@ def add_investor():
         first_name = request.form.get("first_name")
         last_name = request.form.get("last_name")
 
-        error = validate_field(first_name, "First name", "First name cannot be empty.", "admin.add_investor")
-        if error:
+        if error := validate_field(first_name, "First name", "First name cannot be empty.", "admin.add_investor"):
             return error
 
-        error = validate_field(last_name, "Last name", "Last name cannot be empty.", "admin.add_investor")
-        if error:
+        if error := validate_field(last_name, "Last name", "Last name cannot be empty.", "admin.add_investor"):
             return error
 
         email = request.form.get("email")
         if email and not re.match(r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$", email):
             return redirect(
-                url_for("admin.add_investor", _external=False, **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status())
+                url_for(
+                    "admin.add_investor", _external=False, **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status()
+                )
             )
 
         existing_investor = Investor.get_by_email(email=email)  # type: ignore
         if existing_investor:
             return redirect(
                 url_for(
-                    "admin.add_investor", _external=False, **Status(StatusType.ERROR, "Email already exists.").get_status()
+                    "admin.add_investor",
+                    _external=False,
+                    **Status(StatusType.ERROR, AUTH_EMAIL_USED).get_status(),
                 )
             )
 
@@ -187,7 +186,7 @@ def add_investor():
         n_investments = int(request.form.get("n_investments", 0) or 0)
         n_exits = int(request.form.get("n_exits", 0) or 0)
         min_investment = int(request.form.get("min_investment", 0) or 0)
-        max_investment = int(request.form.get("max_investment", 0)or 0)
+        max_investment = int(request.form.get("max_investment", 0) or 0)
 
         selected_rounds = [Round.get_by_id(int(rid)) for rid in selected_round_ids if rid.isdigit()]
         selected_industries = [Industry.get_by_id(int(iid)) for iid in selected_industry_ids if iid.isdigit()]
@@ -220,7 +219,9 @@ def add_investor():
     rounds = Round.get_all()
     industries = Industry.get_all()
 
-    return render_template("admin/add_investor.html", rounds=rounds, industries=industries, status_type=status_type, msg=msg)
+    return render_template(
+        "admin/add_investor.html", rounds=rounds, industries=industries, status_type=status_type, msg=msg
+    )
 
 
 @admin.route("/investor/edit/<int:investor_id>", methods=["GET", "POST"])
@@ -248,14 +249,18 @@ def edit_investor(investor_id):
         email = request.form.get("email")
         if email and not re.match(r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$", email):
             return redirect(
-                url_for("admin.edit_investor", _external=False, **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status())
+                url_for(
+                    "admin.edit_investor", _external=False, **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status()
+                )
             )
 
         existing_investor = Investor.query.filter(Investor.email == email, Investor.id != investor_id).first()
         if existing_investor:
             return redirect(
                 url_for(
-                    "admin.edit_investor", _external=False, **Status(StatusType.ERROR, "Email already exists.").get_status()
+                    "admin.edit_investor",
+                    _external=False,
+                    **Status(StatusType.ERROR, "Email already exists.").get_status(),
                 )
             )
 
@@ -302,7 +307,14 @@ def edit_investor(investor_id):
     rounds = Round.get_all()
     industries = Industry.get_all()
 
-    return render_template("admin/edit_investor.html", investor=investor, rounds=rounds, industries=industries, status_type=status_type, msg=msg)
+    return render_template(
+        "admin/edit_investor.html",
+        investor=investor,
+        rounds=rounds,
+        industries=industries,
+        status_type=status_type,
+        msg=msg,
+    )
 
 
 @admin.route("/investor/delete/<int:investor_id>", methods=["POST"])
@@ -412,14 +424,20 @@ def add_investment_firm():
         email = request.form.get("email")
         if email and not re.match(r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$", email):
             return redirect(
-                url_for("admin.add_investment_firm", _external=False, **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status())
+                url_for(
+                    "admin.add_investment_firm",
+                    _external=False,
+                    **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status(),
+                )
             )
 
         existing_investor = InvestmentFirm.get_by_email(email=email)  # type: ignore
         if existing_investor:
             return redirect(
                 url_for(
-                    "admin.add_investment_firm", _external=False, **Status(StatusType.ERROR, "Email already exists.").get_status()
+                    "admin.add_investment_firm",
+                    _external=False,
+                    **Status(StatusType.ERROR, "Email already exists.").get_status(),
                 )
             )
 
@@ -461,7 +479,9 @@ def add_investment_firm():
     rounds = Round.get_all()
     industries = Industry.get_all()
 
-    return render_template("admin/add_investment_firm.html", rounds=rounds, industries=industries,status_type=status_type, msg=msg)
+    return render_template(
+        "admin/add_investment_firm.html", rounds=rounds, industries=industries, status_type=status_type, msg=msg
+    )
 
 
 @admin.route("/investment-firm/edit/<int:investment_firm_id>", methods=["GET", "POST"])
@@ -472,7 +492,10 @@ def edit_investment_firm(investment_firm_id):
     Need to add validation for phone_number
     """
     status_type, msg = None, None
-    investment_firm = InvestmentFirm.query.get_or_404(investment_firm_id)
+    investment_firm = InvestmentFirm.get_by_id(investment_firm_id)
+    if not investment_firm:
+        abort(404)
+
     if request.method == "POST":
         name = request.form.get("name")
         error = validate_field(name, "Name", "Name cannot be empty.", "admin.edit_investment_firm")
@@ -482,7 +505,11 @@ def edit_investment_firm(investment_firm_id):
         email = request.form.get("email")
         if email and not re.match(r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$", email):
             return redirect(
-                url_for("admin.edit_investment_firm", _external=False, **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status())
+                url_for(
+                    "admin.edit_investment_firm",
+                    _external=False,
+                    **Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status(),
+                )
             )
 
         existing_investor = InvestmentFirm.query.filter(
@@ -491,7 +518,9 @@ def edit_investment_firm(investment_firm_id):
         if existing_investor:
             return redirect(
                 url_for(
-                    "admin.edit_investment_firm", _external=False, **Status(StatusType.ERROR, "Email already exists.").get_status()
+                    "admin.edit_investment_firm",
+                    _external=False,
+                    **Status(StatusType.ERROR, "Email already exists.").get_status(),
                 )
             )
 
@@ -534,13 +563,20 @@ def edit_investment_firm(investment_firm_id):
     industries = Industry.get_all()
 
     return render_template(
-        "admin/edit_investment_firm.html", investment_firm=investment_firm, rounds=rounds, industries=industries,status_type=status_type, msg=msg
+        "admin/edit_investment_firm.html",
+        investment_firm=investment_firm,
+        rounds=rounds,
+        industries=industries,
+        status_type=status_type,
+        msg=msg,
     )
 
 
 @admin.route("/investment-firm/delete/<int:investment_firm_id>", methods=["POST"])
 def delete_investment_firm(investment_firm_id):
-    investment_firm = InvestmentFirm.query.get_or_404(investment_firm_id)
+    investment_firm = InvestmentFirm.get_by_id(investment_firm_id)
+    if not investment_firm:
+        abort(404)
 
     db.session.delete(investment_firm)
     db.session.commit()
@@ -556,9 +592,9 @@ def get_all_users():
     """
     Need to make this page more beautiful
     """
-    users = UserInfo.get_all()
+    user_list = UserInfo.get_all()
 
-    return render_template("admin/get_users.html", users=users)
+    return render_template("admin/get_users.html", users=user_list)
 
 
 @admin.route("/user/add", methods=["GET", "POST"])
@@ -604,24 +640,22 @@ def add_user():
             oauth_provider=OauthProvider.REGULAR,
             is_verified=bool(request.form.get("is_verified")),
             is_admin=bool(request.form.get("is_admin")),
-            )
+        )
         new_user.password = password
+
         db.session.add(new_user)
         db.session.commit()
 
         first_name = request.form.get("first_name")
-        error = validate_field(first_name, "First name", "First name cannot be empty.", "admin.add_user")
-        if error:
+        if error := validate_field(first_name, "First name", "First name cannot be empty.", "admin.add_user"):
             return error
 
         last_name = request.form.get("last_name")
-        error = validate_field(last_name, "Last name", "Last name cannot be empty.", "admin.add_user")
-        if error:
+        if error := validate_field(last_name, "Last name", "Last name cannot be empty.", "admin.add_user"):
             return error
 
         username = request.form.get("username")
-        error = validate_field(username, "Username", "Username cannot be empty.", "admin.add_user")
-        if error:
+        if error := validate_field(username, "Username", "Username cannot be empty.", "admin.add_user"):
             return error
 
         new_user_info = UserInfo(
@@ -637,25 +671,22 @@ def add_user():
             language=request.form.get("language"),
         )
 
-        db.session.add(new_user_info)
-        db.session.commit()
-
-        created_str = request.form.get("created")
-        expires_at_str = request.form.get("expires_at")
-
-        created = datetime.strptime(created_str, "%Y-%m-%d") if created_str else None
-        expires_at = datetime.strptime(expires_at_str, "%Y-%m-%d") if expires_at_str else None
-
         new_user_payment = UserPayment(
             user_id=new_user.id,
             customer_id=request.form.get("customer_id"),
             subscription_id=request.form.get("subscription_id"),
-            created=created,
-            expires_at=expires_at,
+            created=datetime.strptime(created_str, "%Y-%m-%d")
+            if (created_str := request.form.get("created"))
+            else None,
+            expires_at=(
+                datetime.strptime(expires_at_str, "%Y-%m-%d")
+                if (expires_at_str := request.form.get("expires_at"))
+                else None
+            ),
             is_active=bool(request.form.get("is_active")),
         )
 
-        db.session.add(new_user_payment)
+        db.session.add_all((new_user_payment, new_user_info))
         db.session.commit()
         return redirect(url_for("admin.get_all_users"))
     return render_template("admin/add_user.html", languages=language_list, status_type=status_type, msg=msg)
@@ -671,12 +702,30 @@ def edit_user(user_id):
     Need to add style for date fields
     """
     status_type, msg = None, None
-    user = User.query.get_or_404(user_id)
+    user = User.get_by_id(user_id)
+    if not user:
+        abort(404)
+
     user_info = UserInfo.get_by_user_id(user_id)
+    if not user_info:
+        abort(404)
     user_payment = UserPayment.get_by_user_id(user_id)
-    print(user_info)
+    if not user_payment:
+        abort(404)
 
     if request.method == "POST":
+        if not user:
+            status = Status(StatusType.ERROR, "User does not exist.").get_status()
+            return redirect(url_for("admin.edit_user", _external=False, **status))
+
+        if not user_info:
+            status = Status(StatusType.ERROR, "User info does not exist.").get_status()
+            return redirect(url_for("admin.edit_user", _external=False, **status))
+
+        if not user_payment:
+            status = Status(StatusType.ERROR, "User payment does not exist.").get_status()
+            return redirect(url_for("admin.edit_user", _external=False, **status))
+
         email = request.form.get("email")
         password = request.form.get("password")
         confirm_password = request.form.get("confirm_password")
@@ -689,37 +738,24 @@ def edit_user(user_id):
             status = Status(StatusType.ERROR, AUTH_INVALID_EMAIL).get_status()
             return redirect(url_for("auth.edit_user", _external=False, **status))
 
-        existing_user = User.query.filter(User.email == email, User.id != user_id).first()
-        if existing_user:
-            return redirect(
-                url_for(
-                    "admin.edit_user", _external=False, **Status(StatusType.ERROR, "Email already exists.").get_status()
-                )
-            )
+        if User.query.filter(User.email == email, User.id != user_id).first():
+            status = Status(StatusType.ERROR, AUTH_EMAIL_USED).get_status()
+            return redirect(url_for("admin.edit_user", _external=False, **status))
 
         if password != confirm_password:
             status = Status(StatusType.ERROR, AUTH_MISMATCHED_PASSWORDS).get_status()
             return redirect(url_for("admin.edit_user", _external=False, **status))
 
-        user.email = email # type: ignore
-        user.password = password # type: ignore
-        user.is_verified = bool(request.form.get("is_verified"))
-        user.is_admin = bool(request.form.get("is_admin"))
-        db.session.commit()
-
         first_name = request.form.get("first_name")
-        error = validate_field(first_name, "First name", "First name cannot be empty.", "admin.edit_user")
-        if error:
+        if error := validate_field(first_name, "First name", "First name cannot be empty.", "admin.edit_user"):
             return error
 
         last_name = request.form.get("last_name")
-        error = validate_field(last_name, "Last name", "Last name cannot be empty.", "admin.edit_user")
-        if error:
+        if error := validate_field(last_name, "Last name", "Last name cannot be empty.", "admin.edit_user"):
             return error
 
         username = request.form.get("username")
-        error = validate_field(username, "Username", "Username cannot be empty.", "admin.edit_user")
-        if error:
+        if error := validate_field(username, "Username", "Username cannot be empty.", "admin.edit_user"):
             return error
 
         # if pfp := request.files["pfp"]:
@@ -731,6 +767,11 @@ def edit_user(user_id):
         #     except Exception as e:
         #         print(f"An error occurred: {e}")
 
+        user.email = email
+        user.password = password
+        user.is_verified = bool(request.form.get("is_verified"))
+        user.is_admin = bool(request.form.get("is_admin"))
+
         user_info.first_name = first_name
         user_info.last_name = last_name
         user_info.username = username
@@ -739,27 +780,33 @@ def edit_user(user_id):
         user_info.instagram = request.form.get("instagram")
         user_info.twitter = request.form.get("twitter")
         user_info.is_complete = bool(request.form.get("is_complete"))
-        user_info.language = request.form.get("language")
-
-        db.session.commit()
-
-        created_str = request.form.get("created")
-        expires_at_str = request.form.get("expires_at")
-
-        created = datetime.strptime(created_str, "%Y-%m-%d") if created_str else None
-        expires_at = datetime.strptime(expires_at_str, "%Y-%m-%d") if expires_at_str else None
-
+        user_info.language = request.form.get("language", "English")
 
         user_payment.customer_id = request.form.get("customer_id")
         user_payment.subscription_id = request.form.get("subscription_id")
-        user_payment.created = created
-        user_payment.expires_at = expires_at
+        user_payment.created = (
+            datetime.strptime(created_str, "%Y-%m-%d") if (created_str := request.form.get("created")) else None
+        )
+        user_payment.expires_at = (
+            datetime.strptime(expires_at_str, "%Y-%m-%d")
+            if (expires_at_str := request.form.get("expires_at"))
+            else None
+        )
         user_payment.is_active = bool(request.form.get("is_active"))
 
         db.session.commit()
 
         return redirect(url_for("admin.get_all_users"))
-    return render_template("admin/edit_user.html", user=user, user_info=user_info, user_payment=user_payment, languages=language_list, status_type=status_type, msg=msg)
+
+    return render_template(
+        "admin/edit_user.html",
+        user=user,
+        user_info=user_info,
+        user_payment=user_payment,
+        languages=language_list,
+        status_type=status_type,
+        msg=msg,
+    )
 
 
 @admin.route("/user/delete/<int:user_id>", methods=["POST"])
